Add logging around missing refresh_tokens.
It's possible for a user to do the full 3LO dance and not receive a refresh
token. This is usually unexpected (such as the case of a server requesting a
token for a user who already has one for that same client ID), and can be hard
to diagnose.

This just adds a log message, with a pointer to the necessary argument to
force a new refresh token.