Clean up test envs.
We have a number of extra envs due to checking two versions of pyopenssl (0.13
and 0.14); this isn't actually necessary, since 0.14 is the one we really care
about for the external repo.

This cleans up the cases and names (and drastically shortens `tox.ini`), but
keeps the old envs available for manual testing.