Allow the config dir to be missing for read.
As part of #162, we added a check that the directory containing the well-known
file exists. In the case that we're about to write the file, this nicely
solves the issue raised in #151. However, in the case that we're entirely
indifferent to the existence of the well-known file, this is an overzealous
check.

This PR simply tweaks the code to only fail in the case that the directory
doesn't exist when writing the well-known file.